Separate utility classes. Rewrite URL encoder to split path names and encode path components only.
