+++ src/main/java/synapticloop/b2/request/B2DeleteFileVersionRequest.java
@@ -24,7 +24,7 @@ import org.apache.http.util.EntityUtils;
 import synapticloop.b2.exception.B2ApiException;
 import synapticloop.b2.response.B2AuthorizeAccountResponse;
 import synapticloop.b2.response.B2DeleteFileVersionResponse;
-import synapticloop.b2.util.Helper;
+import synapticloop.b2.util.URLEncoder;
 
 /**
  * <p>Deletes one version of a file from B2.</p>
@@ -57,7 +57,7 @@ public class B2DeleteFileVersionRequest extends BaseB2Request {
 	public B2DeleteFileVersionRequest(CloseableHttpClient client, B2AuthorizeAccountResponse b2AuthorizeAccountResponse, String fileName, String fileId) {
 		super(client, b2AuthorizeAccountResponse, b2AuthorizeAccountResponse.getApiUrl() + B2_DELETE_FILE_VERSION);
 
-		this.addProperty(B2RequestProperties.KEY_FILE_NAME, Helper.urlEncodeFileName(fileName));
+		this.addProperty(B2RequestProperties.KEY_FILE_NAME, URLEncoder.encode(fileName));
 		this.addProperty(B2RequestProperties.KEY_FILE_ID, fileId);
 	}
